# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  models.CreateTodoListRequest:
    properties:
      description:
        maxLength: 500
        type: string
      name:
        maxLength: 100
        minLength: 1
        type: string
    required:
    - name
    type: object
...
  models.ErrorResponse:
    properties:
      code:
        type: string
      details:
        additionalProperties: true
        type: object
      message:
        type: string
    type: object
...
  models.PaginatedListsResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/models.TodoList'
        type: array
      pagination:
        $ref: '#/definitions/models.Pagination'
    type: object
  models.Pagination:
    properties:
      limit:
        type: integer
      page:
        type: integer
      totalItems:
        type: integer
      totalPages:
        type: integer
    type: object
...
  models.TodoList:
    properties:
      createdAt:
        type: string
      description:
        maxLength: 500
        type: string
      id:
        type: string
      name:
        maxLength: 100
        minLength: 1
        type: string
      todoCount:
        type: integer
      updatedAt:
        type: string
      userId:
        type: string
    required:
    - name
    type: object
...
  models.UpdateTodoListRequest:
    properties:
      description:
        maxLength: 500
        type: string
      name:
        maxLength: 100
        minLength: 1
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@todolist-api.com
    name: API Support
  description: A RESTful API for managing todo lists and tasks with user authentication
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: TodoList API
  version: "1.0"
paths:
...
  /lists:
    get:
      description: Get all todo lists for the authenticated user with pagination
      parameters:
      - description: 'Page number (default: 1)'
        in: query
        name: page
        type: integer
      - description: 'Items per page (default: 20, max: 100)'
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.PaginatedListsResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - Bearer: []
      summary: Get all todo lists
      tags:
      - Lists
    post:
      consumes:
      - application/json
      description: Create a new todo list for the authenticated user
      parameters:
      - description: List details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.CreateTodoListRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.TodoList'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - Bearer: []
      summary: Create a new todo list
      tags:
      - Lists
  /lists/{listId}:
    delete:
      description: Delete a specific todo list and all its todos
      parameters:
      - description: List ID (UUID)
        in: path
        name: listId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - Bearer: []
      summary: Delete a todo list
      tags:
      - Lists
    get:
      description: Get a specific todo list by ID
      parameters:
      - description: List ID (UUID)
        in: path
        name: listId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.TodoList'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - Bearer: []
      summary: Get a specific todo list
      tags:
      - Lists
    put:
      consumes:
      - application/json
      description: Update a specific todo list
      parameters:
      - description: List ID (UUID)
        in: path
        name: listId
        required: true
        type: string
      - description: Updated list details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.UpdateTodoListRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.TodoList'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - Bearer: []
      summary: Update a todo list
      tags:
      - Lists
...
securityDefinitions:
  Bearer:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...